Refactor file structure and update import paths for template validation
